% Generated by roxygen2 (4.0.1): do not edit by hand
\name{decomposeTrait}
\alias{decomposeTrait}
\title{Decompose trait values according to the model z=G+I+E+epsilon}
\usage{
decomposeTrait(data, GEValues)
}
\arguments{
\item{data}{a data.table returned by extractPop}

\item{GEValues}{a matrix of GEValues}
}
\description{
Decompose trait values according to the model z=G+I+E+epsilon
}

