% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estimBeta}
\alias{estimBeta}
\title{Regression slope of recipient on donor values in an epidemic}
\usage{
estimBeta(epidemic, GEValues, atInfection = FALSE, sampledOnly = TRUE,
  tinfMin = 0, tinfMax = Inf, report = FALSE)
}
\description{
Regression slope of recipient on donor values in an epidemic
}

