% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estimH2}
\alias{estimH2}
\title{Broad-sense heritability of a pathogen trait}
\usage{
estimH2(epidemic, GEValues, sampledOnly = TRUE, activeOnly = FALSE,
  tMin = 0, tMax = Inf, lastN = Inf)
}
\arguments{
\item{epidemic}{a list of objects returned from simulateEpidemic}

\item{GEValues}{genotype-environment trait values}

\item{sampledOnly}{logical, indicating if only sampled individuals should be included in the heritability calculation}

\item{tMin,tMax}{numeric, time interval for which to measure the heritability. if sampledOnly is TRUE this would be
the times of sampling for the individuals; otherwise, this would be the times of infection.}
}
\value{
numeric indicating the estimated heritability for infected individuals in the population.
}
\description{
Broad-sense heritability of a pathogen trait
}

