% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estimH2aov}
\alias{estimH2aov}
\title{Estimating broad-sense heritability through ANOVA}
\usage{
estimH2aov(epidemic = NULL, data = NULL, GEValues, by = list("gene"),
  sampledOnly = TRUE, activeOnly = FALSE, tMin = 0, tMax = Inf,
  lastN = Inf, report = FALSE)
}
\arguments{
\item{epidemic}{a list of objects returned from simulateEpidemic}

\item{NULL}{or a data.table such as the element gen in an epidemic list}

\item{GEValues}{genotype-environment trait values}

\item{by}{a character string which can be evaluated as expression in the by clause of data.table}

\item{sampledOnly}{logical, indicating if only sampled individuals should be included in the heritability calculation}

\item{tMin,tMax}{numeric, time interval for which to measure the heritability. if sampledOnly is TRUE this would be
the times of sampling for the individuals; otherwise, this would be the times of infection.}
}
\description{
Estimating broad-sense heritability through ANOVA
}

