% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractDRCouples}
\alias{extractDRCouples}
\title{Extract donor-recipient couples from an epidemic}
\usage{
extractDRCouples(epidemic, sampledOnly = TRUE, activeOnly = FALSE,
  tMin = 0, tMax = Inf)
}
\arguments{
\item{epidemic}{a list representing an epidemic (as returned by simulateEpidemic)}

\item{sampledOnly}{a logical indicating whether only recovered individuals should be extracted}

\item{activeOnly}{a logical indicating whether only currently infected and alive individuals should be extracted}

\item{tMin,tMax}{a numericals indicating the interval of observation times for extracted individuals. If sampledOnly is
TRUE the observation time is the time of sampling (recovery), otherwise it is the time of infection. Default: tMin=0, tMax=Inf.}
}
\value{
a data.table with rows corresponding to the extracted couples with the following columns:
idD: id of the donor
id: id of the recipient
envd: environment type of the donor
gd: transmitted strain from the donor to the recipient
ed: donor special environmental effect at the moment of transmission
env: environment type of the recipient
gene: strain in the recipient at the moment of sampling or at the current moment
e: special environmental effects of the recipient
tauR: age of infection in the recipient (in timeStep units)
tauDAtInf: age of infection in the donor at the moment of transmission (in timeStep units)
tauD: time (in timeStep units) in the donor from the moment of transmission until the moment of sampling or the current moment
taum: =tauD+tauR
eD: special environmental effect in the donor at the moment of sampling
gD: strain in the donor at the moment of sampling
}
\description{
Extract donor-recipient couples from an epidemic
}

