% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractPop}
\alias{extractPop}
\title{Extract a sub-population of individuals involved in an epidemic}
\usage{
extractPop(epidemic, sampledOnly = TRUE, activeOnly = FALSE, tMin = 0,
  tMax = Inf, lastN = Inf)
}
\arguments{
\item{epidemic}{a list representing an epidemic (as returned by simulateEpidemic)}

\item{sampledOnly}{a logical indicating whether only recovered individuals should be extracted}

\item{activeOnly}{a logical indicating whether only currently infected and alive individuals should be extracted}

\item{tMin,tMax}{a numericals indicating the interval of observation times for extracted individuals. If sampledOnly is
TRUE the observation time is the time of sampling (recovery), otherwise it is the time of infection. Default: tMin=0, tMax=Inf.}

\item{lastN}{numeric indicating whether to return only the most recently observed lastN individuals (default is Inf)}
}
\value{
a data.table with rows corresponding to the extracted individuals
}
\description{
Extract a sub-population of individuals involved in an epidemic
}

