% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractPP}
\alias{extractPP}
\title{Extract phylogenetic pairs from a phylogeny}
\usage{
extractPP(tree, threshold = Inf, firstN = length(tree$tip.label))
}
\arguments{
\item{tree}{a phylo object}

\item{threshold}{numeric indicating the maximum pair distance that should be allowed in the returned pairs}

\item{firstN}{integer indicating whether only the nearest firstN pairs should be returned}
}
\value{
a data.table with four columns:
i, j : the members of each phylogenetic pair. For each entry (i,j) a symmetric entry (j,i) is present
d: the phylogenetic distance between i and j
idPair: equal to min(i,j) for each entry - the identifier of each pair.
}
\description{
Extract phylogenetic pairs from a phylogeny
}

