% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estimH2}
\alias{estimH2}
\title{Broad-sense heritability of a pathogen trait}
\usage{
estimH2(epidemic, GEValues, sampledOnly = TRUE, tinfMin = 0,
  tinfMax = Inf)
}
\arguments{
\item{epidemic}{a list of objects returned from simulateEpidemic}

\item{GEValues}{genotype-environment trait values}

\item{sampledOnly}{logical, indicating if only sampled individuals should be included in the heritability calculation}

\item{tinfMin,tinfMax}{numeric, time interval for which to measure the heritability}
}
\value{
numeric indicating the estimated heritability for infected individuals in the population.
}
\description{
Broad-sense heritability of a pathogen trait
}

