% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estimH2}
\alias{estimH2}
\title{Broad-sense heritability of a pathogen trait}
\usage{
estimH2(epidemic = NULL, data = NULL, GEValues = NULL,
  sampledOnly = TRUE, activeOnly = FALSE, tMin = 0, tMax = Inf,
  firstN = Inf, lastN = Inf)
}
\arguments{
\item{epidemic}{a list of objects returned from simulateEpidemic}

\item{data}{NULL or a data.table such as the element gen in an epidemic list}

\item{GEValues}{genotype-environment trait values for calculating the z-values; if NULL it is assumed that the
z-values are already in data, or epidemic$gen.}

\item{sampledOnly,activeOnly,tMin,tMax,firstN,lastN}{parameters passed to extractPop}
}
\value{
numeric indicating the estimated heritability for infected individuals in the population.
}
\description{
Broad-sense heritability of a pathogen trait
}

