% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estimH2aov}
\alias{estimH2aov}
\title{Estimating intraclass correlation (ICC) through ANOVA}
\usage{
estimH2aov(epidemic = NULL, data = NULL, GEValues = NULL,
  sampledOnly = TRUE, activeOnly = FALSE, tMin = 0, tMax = Inf,
  firstN = Inf, lastN = Inf, by = list("gene"), report = FALSE)
}
\arguments{
\item{epidemic}{a list of objects returned from simulateEpidemic}

\item{data}{NULL or a data.table such as the element gen in an epidemic list}

\item{GEValues}{genotype-environment trait values for calculating the z-values; if NULL it is assumed that the
z-values are already in data, or epidemic$gen.}

\item{sampledOnly,activeOnly,tMin,tMax,firstN,lastN}{parameters passed to extractPop}

\item{by}{a character string which can be evaluated as expression in the by clause of data.table
the times of sampling for the individuals; otherwise, this would be the times of infection.}

\item{report}{a logical indicating what result should be returned. If FALSE, only the ICC value is returned,
otherwise a list of statistics from the ANOVA calculation}
}
\value{
See report parameter
}
\description{
Estimating intraclass correlation (ICC) through ANOVA
}

