% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateContinuousTime.R
\name{extractPP}
\alias{extractPP}
\title{Extract phylogenetic pairs from a phylogeny}
\usage{
extractPP(tree, threshold = Inf, firstN = Inf)
}
\arguments{
\item{tree}{a phylo object}

\item{threshold}{numeric indicating the maximum pair distance that should be allowed in the returned pairs}

\item{firstN}{integer indicating whether only the nearest firstN pairs should be returned}
}
\value{
a data.table with four columns:
i, j : integers - the members of each phylogenetic pair. For each entry (i,j) a symmetric entry (j,i) is present; 
   to obtain the corresponding tip labels in the tree use tree$tip.label[i] and tree$tip.label[j] respectively.
d: the phylogenetic distance between i and j
idPair: equal to min(i,j) for each entry - the identifier of each pair.
}
\description{
Extract phylogenetic pairs from a phylogeny
}

