% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeTree}
\alias{makeTree}
\title{Transmission chain between sampled individuals in an epidemic}
\usage{
makeTree(epideimic)
}
\arguments{
\item{epidemic}{list returned by simulateEpidemic}
}
\description{
Transmission chain between sampled individuals in an epidemic
}

