% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeTree}
\alias{makeTree}
\title{Transmission chain between sampled individuals in an epidemic}
\usage{
makeTree(epidemic, tips = NULL, idTips = NULL)
}
\arguments{
\item{epidemic}{list returned by simulateEpidemic}

\item{tips,idTips}{integer vectors (see details)}
}
\description{
Transmission chain between sampled individuals in an epidemic
}
\details{
By default all individuals sampled during the active phase of the epidemic are
included in the phylogeny. If specified, the argument tips has priority to the arugment
idTips. The argument tips is sorted in increasing order and specifies the tip indices
(numbers in the range 1:epidemic$nTips) to be retained in the phylogeny. If NULL (default),
the argument idTips is used to identify which individuals to be included in the phylogeny.
This is useful if, for example, one has searched all individuals who were sampled during
a specified time interval of the epidemic.
}

