% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateContinuousTime.R
\name{rateTransTemplate_32}
\alias{rateTransTemplate_32}
\title{state transition template matrix (main diagonal omitted)
element [i, j] is a multiplier for the rate of transition 
from state i to state j+k, where k=1 if j>=i and k=0 otherwise}
\usage{
rateTransTemplate_32()
}
\description{
state transition template matrix (main diagonal omitted)
element [i, j] is a multiplier for the rate of transition 
from state i to state j+k, where k=1 if j>=i and k=0 otherwise
}

